Trying something new
Moved _all_ iwyu files into a special subdirectory, since leaving them
in the typical `include` and `lib` directories ends up with those files
being linked in folders in `/usr/local`. Since iwyu is what needs these
files (they're not meant to be external), that doesn't make sense, and
in fact can interfere with other packages.

The binaries are really all that's necessary, so symlinks are made to
them after everything else is in its place (the actual binaries must
still be in the right place relative to the lib and include dirs, so
there's a bin directory inside the special one in addition to a typical
bin directory at the root of the package directory).