Upgrade to iwyu 0.5
Upgrade to the latest include-what-you-use. The latest release bundles
needed C headers in the actual archive, which is nice, but without the
C++ headers I can't yet get rid of the Xcode dependency. So I've been
forced to bump it to 7.0.

On the upside, I figured out how to remove the default OS X build (now
failing because it defaults to Xcode 6.1).